Quieting compiler warnings.  NOTE: these raise my eyebrows as those warnings may well be valid.  Vector.size() and company return vector<type>::size_t, which may be bigger than int on 64-bit systems and unsigned.  But the counter variables are defined as int as use negatives as meaningful values, so I cannot just replace member defs with the correct unsigned type.  I think there is definitely room for problems casting down to int.

+++ mscore_kgpu.cpp
@@ -305,7 +305,7 @@ bool mscore_kgpu::playback_sequence() {
     assert_consistency(); // no effect in release build
     restore_mscore_internals(++m_current_playback_sequence);
     m_current_playback_sequence_begin = m_cached_sequences_index[m_current_playback_sequence];
-    m_current_playback_sequence_end = ((m_current_playback_sequence+1)==m_cached_sequences_index.size())?
+    m_current_playback_sequence_end = ((m_current_playback_sequence+1)==(int)m_cached_sequences_index.size())?
         m_cached_sequences_l.size():
         m_cached_sequences_index[m_current_playback_sequence+1];
    int len = (m_current_playback_sequence_end-m_current_playback_sequence_begin);
